Trigger Maven actions in subprojects
If the main project doesn't have an execution defined for the requested
goal, trigger the action in all subprojects. If none of them has a
corresponding execution, fall back to running the action without checks
in the selected project.

This version has problems if multiple launches are triggered at the
same time, as only one of them is actually executed in that case.