Fix selection of archetype radio buttons
With the old solution, it was possible to select multiple radio
buttons in the new Vaadin Maven project wizard by clicking on the
text labels, as SWT radio button handling was not invoked.

There doesn't seem to be a clean way to invoke it from the outside,
so iterating over all radio buttons.